check posgis version

+++ test/integration/data-types.test.js
@@ -276,45 +276,56 @@ describe(Support.getTestDialectTeaser('DataTypes'), function() {
   });
 
   it('should parse an empty GEOMETRY field', function () {
-   var Type = new Sequelize.GEOMETRY();
+    var Type = new Sequelize.GEOMETRY();
 
-   // MySQL 5.7 or above doesn't support POINT EMPTY
-   if (dialect === 'mysql' && semver.gte(current.options.databaseVersion, '5.7.0')) {
-     return;
-   }
+    // MySQL 5.7 or above doesn't support POINT EMPTY
+    if (dialect === 'mysql' && semver.gte(current.options.databaseVersion, '5.7.0')) {
+      return;
+    }
 
-   // Postgres 9.5.0 or above doesn't support POINT EMPTY
-   if (/^postgres/.test(dialect) && semver.gte(current.options.databaseVersion, '9.5.0')) {
-     return;
-   }
+    return new Sequelize.Promise((resolve, reject) => {
+      if (/^postgres/.test(dialect)) {
+        current.query(`SELECT extversion FROM pg_catalog.pg_extension WHERE extname='postgis';`)
+            .then((result) => {
+              if (result[0][0] && semver.lte(result[0][0].extversion, '2.1.7')) {
+                console.log(result[0][0]); //ya I know :)
+                resolve(true);
+              } else {
+                resolve();
+              }
+            }).catch(reject);
+      } else {
+        resolve(true);
+      }
+    }).then((runTests) => {
+      if (current.dialect.supports.GEOMETRY && runTests) {
+        current.refreshTypes();
 
-   if (current.dialect.supports.GEOMETRY) {
-     current.refreshTypes();
+        var User = current.define('user', { field: Type }, { timestamps: false });
+        var point = { type: "Point", coordinates: [] };
 
-     var User = current.define('user', { field: Type }, { timestamps: false });
-     var point = { type: "Point", coordinates: [] };
-
-     return current.sync({ force: true }).then(function () {
-       return User.create({
-          //insert a null GEOMETRY type
-          field: point
+        return current.sync({ force: true }).then(function () {
+          return User.create({
+            //insert a null GEOMETRY type
+            field: point
+          });
+        }).then(function () {
+          //This case throw unhandled exception
+          return User.findAll();
+        }).then(function(users){
+          if (dialect === 'mysql') {
+            // MySQL will return NULL, becuase they lack EMPTY geometry data support.
+            expect(users[0].field).to.be.eql(null);
+          } else if (dialect === 'postgres' || dialect === 'postgres-native') {
+            //Empty Geometry data [0,0] as per https://trac.osgeo.org/postgis/ticket/1996
+            expect(users[0].field).to.be.deep.eql({ type: "Point", coordinates: [0,0] });
+          } else {
+            expect(users[0].field).to.be.deep.eql(point);
+          }
         });
-      }).then(function () {
-        //This case throw unhandled exception
-        return User.findAll();
-      }).then(function(users){
-        if (dialect === 'mysql') {
-          // MySQL will return NULL, becuase they lack EMPTY geometry data support.
-          expect(users[0].field).to.be.eql(null);
-        } else if (dialect === 'postgres' || dialect === 'postgres-native') {
-          //Empty Geometry data [0,0] as per https://trac.osgeo.org/postgis/ticket/1996
-          expect(users[0].field).to.be.deep.eql({ type: "Point", coordinates: [0,0] });
-        } else {
-          expect(users[0].field).to.be.deep.eql(point);
-        }
-      });
-   }
- });
+      }
+    });
+  });
 
   if (dialect === 'postgres' || dialect === 'sqlite') {
     // postgres actively supports IEEE floating point literals, and sqlite doesn't care what we throw at it
